backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.6.0
redis==5.0.6
celery==5.4.0
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4
pydantic==1.10.13
sentence-transformers>=2.2.0
torch>=2.0.0
faiss-cpu>=1.7.4

